API/requirements.txt
fastapi==0.62.0
uvicorn==0.13.0
numpy==1.18.5
pandas==1.1.4
scipy==1.4.1
google-cloud-bigquery==2.6.0
google-cloud-bigquery-storage==2.1.0 
pyarrow==2.0.0
